Intercept android_dlopen_ext to handle namespace.
Android N come with support for putting native libraries in different
namespace. If a.so is opened with some namespace, when a.so open b.so
with dlopen, Android dlopen know it's called from a.so and then will
open b.so in same namespace, but since we intercept dlopen already,
that means Android dlopen thinks it's called from egltrace.so and
then open b.so in wrong namespace. To work around this, we record
the namespace for a.so when opening a.so and use android_dlopen_ext to
explicitly request open b.so in the recorded name space.